+++ packages/node/src/ethereum/api.service.ethereum.test.ts
@@ -83,4 +83,12 @@ describe('ApiService', () => {
     console.log('Finalized height', height);
     expect(height).toBeGreaterThan(16_000_000);
   });
+
+  it('ensure api errorCode is exposed when throwing', async () => {
+    expect.assertions(1);
+    return apiService
+      .safeApi(17520376)
+      .getCode('0x75F0398549C9fDEa03BbDde388361827cb376D5')
+      .catch((e) => expect(e.code).toBe('INVALID_ARGUMENT'));
+  });
 });